+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Multiple capabilities in interactive install command
+- Providers priority array
+- Package manager abstraction
 
 ## [0.0.17] - 2021-06-08
 ### Changed